requirements.txt
python-telegram-bot==21.0.1
anthropic==0.40.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.3.0
google-api-python-client==2.116.0
google-analytics-data==0.18.13
requests==2.31.0
feedparser==6.0.11
apscheduler==3.10.4
python-dotenv==1.0.1
openai
pytz
httpx
urllib3<2
